requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
opencv-python
scikit-learn
matplotlib
seaborn
pandas
tqdm
